create CounterfactualLocalImportanceChart
